Refactor - 데이터 처리 컴포넌트를 분리하며 Inner 컴포넌트 호출 거리 최소화
