--- components/brands/BrandList.type.ts
@@ -1,6 +1,0 @@
-import { ContItems } from 'apis/models/Categories.type';
-
-export default interface BrandsProps {
-  name: string;
-  data: ContItems[];
-}
