+++ components/brands/index.type.ts
@@ -1,0 +1,5 @@
+import { ConCategory2s } from 'apis/models/Categories.type';
+
+export default interface BrandsProps {
+  data?: ConCategory2s;
+}
